+++ packages/components/src/custom-select-control-v2/types.ts
@@ -57,6 +57,10 @@ export type _CustomSelectProps = {
 	 * Accessible label for the control.
 	 */
 	label: string;
+	/**
+	 * Optional props passed to the Ariakit Select Popover.
+	 */
+	popoverProps?: Ariakit.SelectPopoverProps;
 };
 
 export type CustomSelectProps = _CustomSelectProps &